requirements.txt
python-dotenv==1.0.1
annotated-types==0.6.0
anyio==4.2.0
blinker==1.7.0
CacheControl==0.14.0
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudevents==1.9.0
cryptography==42.0.2
deprecation==2.1.0
fastapi==0.109.2
firebase-admin==6.4.0
firebase-functions==0.2.0
Flask==3.0.2
Flask-Cors==4.0.0
functions-framework==3.5.0
google-api-core==2.16.2
google-api-python-client==2.116.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.41.0
google-cloud-appengine-logging==1.4.1
google-cloud-audit-log==0.2.5
google-cloud-bigquery==3.17.2
google-cloud-core==2.4.1
google-cloud-firestore==2.14.0
google-cloud-logging==3.9.0
google-cloud-resource-manager==1.12.1
google-cloud-speech==2.24.1
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-events==0.11.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpc-google-iam-v1==0.13.0
grpcio==1.60.1
grpcio-status==1.60.1
gunicorn==21.2.0
h11==0.14.0
httplib2==0.22.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
msgpack==1.0.7
numpy==1.26.4
packaging==23.2
proto-plus==1.23.0
protobuf==4.25.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
shapely==2.0.2
six==1.16.0
sniffio==1.3.0
starlette==0.36.3
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.2.0
uvicorn==0.27.0.post1
watchdog==4.0.0
Werkzeug==3.0.1
